Fix the connect problem when using iOS
Addresses #71

When running against a captive portal, iOS for some reason hammers
/index.html. On comitup-web, this triggers a D-Bus call to comitup,
asking for a list of candidate APs. This also happens immediately after
the connect.html call, which calls D-Bus to start the connection. This
collision apparently causes the connect to fail.

Mitigate by TTL caching the return of the list of AP call, and call this
cached function immediately before the connect request.

Tested with a Pi 4 and a "7th generation" iPad running v14.4.