Remove crash on extended unicoded SSIDs
The SSID selection web page was crashing when seeing available APs
containing extended unicode characters, due to improper decoding in
iwscan.

With this fix, APs with unicode content are visible as choices on the
web page.

Note that there is still a problem with Avahi. Comitup.log reports
"Exception encountered adding avahi record" after the A and AAAA record
logs, and the name does not appear in the Avahi Discover list.

Adresses #204